Add generator for initial positions
A file with initial positions was provided for a 3D simulation but since
we are working in 2D we need different positions. This generator first
generates a packed triangular grid of points, scales it up and moves the
points around, and then confines the grid to a square of smaller size by
randomly repositioning the points from outside the square to inside,
making sure they do not overlap, neither inside the square nor around
the periodic boundary conditions.