Read init data through Python into C++ and verify
This creates a Python module using Boost.Python and defines a function
simulate(). For now the function verifies the type and shape of the
initial position data array and the size of the box, then returns the
array. This module is loaded in Python which also reads the initial
positions file (init2d.dat), and then calls the simulate function.

This commit also includes the CMakeLists.txt and Makefile