Add pbc and return value for Python
This add periodic boundary conditions that are only checked when the
particle is near the boundary. The particle is also moved to the other
side of the box when it crosses a boundary.
After all partices have had a chance to move a frame is saved to an
array. At the end this array of all frames is returned back to Python
which can now make a pretty animation.