Add density scaling during simulation
Changing the density within the same simulation saves on equilibration
time. The results for different densites are returned as one large array
with a specified number of matrices per density. python.py and
python_pretty.py are updated to run the simulation and process it's
data. 3d_hist.py can run the simulation but does not show all or the
correct data yet.