Better Péclet number, logging, init. Add ODR
This sets the simulation parameters such that we should in theory get a
better simulation. It also adds better logging so we know various other
calculated values and can see when a simulation breaks. It also adds a
better algorithm for initial conditions and adds some files it
generated. The histogram is run through ODR to find the tops of two
overlapping Gaussian curves.