Improve efficiency by sorting particles in bins
Before starting the simulation pointers to the particles are stored in a
2d grid of bins. This way we roughly know what particles are close to
what other particles, so we only have to check the interaction force for
those ~72 particles, and not all 512 particles in the system, possibly
multiple times due to pbc. The interaction force formula is also
modified to be aroud 23% faster, and the repulsion_strength parameter
is set to constexpr 1.0 again.
Also improved plotting in python.py.